Stream Enrich: change package name (closes #3340)
