+++ 3-enrich/stream-enrich/project/BuildSettings.scala
@@ -32,7 +32,7 @@ object BuildSettings {
   // Makes our SBT app settings available from within the app
   lazy val scalifySettings = Seq(sourceGenerators in Compile <+= (sourceManaged in Compile, version, name, organization) map { (d, v, n, o) =>
     val file = d / "settings.scala"
-    IO.write(file, """package com.snowplowanalytics.snowplow.enrich.kinesis.generated
+    IO.write(file, """package com.snowplowanalytics.snowplow.enrich.stream.generated
       |object Settings {
       |  val organization = "%s"
       |  val version = "%s"
