+++ 3-enrich/stream-enrich/src/main/scala/com.snowplowanalytics.snowplow.enrich/stream/CredentialsLookup.scala
@@ -16,7 +16,7 @@
  * See the Apache License Version 2.0 for the specific language
  * governing permissions and limitations there under.
  */
-package com.snowplowanalytics.snowplow.enrich.kinesis
+package com.snowplowanalytics.snowplow.enrich.stream
 
 // Amazon
 import com.amazonaws.auth._
