+++ 3-enrich/stream-enrich/src/main/scala/com.snowplowanalytics.snowplow.enrich/stream/sinks/ISink.scala
@@ -16,12 +16,19 @@
  * See the Apache License Version 2.0 for the specific language
  * governing permissions and limitations there under.
  */
-package com.snowplowanalytics.snowplow.enrich.kinesis.sinks
+package com.snowplowanalytics.snowplow.enrich
+package stream.sinks
 
 /**
- * Whether the sink is for good rows or bad rows
+ * An interface for all sinks to use to store events.
  */
-object InputType extends Enumeration {
-  type InputType = Value
-  val Good, Bad = Value
+trait ISink {
+
+  /**
+   * Side-effecting function to store the EnrichedEvent
+   * to the given output stream.
+   */
+  def storeEnrichedEvents(events: List[(String, String)]): Boolean
+
+  def flush(): Unit
 }
