+++ 3-enrich/stream-enrich/src/main/scala/com.snowplowanalytics.snowplow.enrich/stream/sources/StdinSource.scala
@@ -19,7 +19,7 @@
 
 package com.snowplowanalytics
 package snowplow.enrich
-package kinesis
+package stream
 package sources
 
 // Scala
