+++ 3-enrich/stream-enrich/src/main/scala/com.snowplowanalytics.snowplow.enrich/stream/sources/TestSource.scala
@@ -18,7 +18,7 @@
  */
 package com.snowplowanalytics
 package snowplow.enrich
-package kinesis
+package stream
 package sources
 
 // Iglu
