+++ 3-enrich/stream-enrich/src/test/scala/com.snowplowanalytics.snowplow.enrich.stream/bad/InvalidEnrichedEventSpec.scala
@@ -13,7 +13,7 @@
  * governing permissions and limitations there under.
  */
 package com.snowplowanalytics.snowplow
-package enrich.kinesis
+package enrich.stream
 package bad
 
 // Commons Codec
