+++ 3-enrich/stream-enrich/src/test/scala/com.snowplowanalytics.snowplow.enrich.stream/sources/AbstractSourceSpec.scala
@@ -18,7 +18,7 @@
  */
 package com.snowplowanalytics
 package snowplow.enrich
-package kinesis
+package stream
 package sources
 
 // Scalaz
